Day 9/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.18.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
faiss-cpu>=1.7.4
pydantic>=2.5.0
scikit-learn>=1.4.0
numpy>=1.26.0
chromadb>=0.4.0
openai>=1.0.0
matplotlib>=3.7.0
